Feat: add auth router
