--- src/controllers/signOut.controller.js
@@ -1,6 +1,0 @@
-exports.signOut = async function (req, res, next) {
-  res.clearCookie("accessToken");
-  res.clearCookie("refreshToken");
-
-  res.send({ result: "ok", message: "logout successful" });
-};
