+++ src/loaders/routers.js
@@ -1,18 +1,14 @@
-const indexRouter = require("../routes/index");
 const keywordsRouter = require("../routes/keywords");
 const autoCompletionsRouter = require("../routes/autoCompletions");
 const videosRouter = require("../routes/videos");
-const signInRouter = require("../routes/signIn");
-const signOutRouter = require("../routes/signOut");
+const authRouter = require("../routes/auth");
 const adminRouter = require("../routes/admin");
 
 async function routerLoader(app) {
-  app.use("/", indexRouter);
   app.use("/keywords", keywordsRouter);
   app.use("/auto-completions", autoCompletionsRouter);
   app.use("/videos", videosRouter);
-  app.use("/signIn", signInRouter);
-  app.use("/signOut", signOutRouter);
+  app.use("/auth", authRouter);
   app.use("/admin", adminRouter);
 }
