--- src/routes/signOut.js
@@ -1,9 +1,0 @@
-const express = require("express");
-
-const signOutController = require("../controllers/signOut.controller");
-
-const router = express.Router();
-
-router.get("/", signOutController.signOut);
-
-module.exports = router;
